Admin 모든 테이블(클릭시 경로 없는 테이블 제외) 행 호버 선택 반응 적용

+++ src/main/webapp/admin/promo_coupon.jsp
@@ -400,7 +400,7 @@
 										<a href="updateCouponPage.mdo?coupon_code=${CouponInfo.coupon_code}">수정</a>
 									</div>
 									<div class="popup_btn">
-										<a href="deleteCoupon.mdo?coupon_code=${CouponInfo.coupon_code}">삭제</a>
+										<a href="deleteCoupon.mdo?coupon_code=${CouponInfo.coupon_code}" onclick="return confirm('삭제하시겠습니까?');">삭제</a>
 									</div>
 								</div>
 							</form>
